admin-panel/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.6.0
python-multipart==0.0.9
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pydantic[email]==2.6.0
PyJWT==2.8.0